增加注册界面，删除多余界面
1.增加老师和学生的注册界面pages/regis

2.删除多余测试界面
