+++ client/app.json
@@ -1,16 +1,11 @@
 {
   "pages": [
     "pages/test_index/test_index",
+    "pages/regis/regis",
     "pages/course_signhistory/course_signhistory",
     "pages/stu_course/stu_course",
     "pages/stu_index/stu_index",
-    "pages/stu_regis/stu_regis",
     "pages/index/index",
-    "pages/firstIndex/firstIndex",
-    "pages/tea_regis/tea_regis",
-    "pages/index1/index",
-    "pages/addCgi/addCgi",
-    "pages/tea_index/tea_index",
     "pages/course_announce/course_announce"
   ],
   "window": {
