--- client/pages/addCgi/addCgi.json
@@ -1,3 +1,0 @@
-{
-    "navigationBarTitleText": "快速添加 CGI"
-}
