--- client/pages/firstIndex/1
@@ -1,4 +1,0 @@
-字体：
-19 迷你简体稚艺
-361 逐浪创意流珠体
-367 逐浪小雪钢笔体
