--- client/pages/firstIndex/firstIndex.wxml
@@ -1,33 +1,0 @@
-<!--index.wxml-->
-<!--
-<view class="container">
-  <view class="userinfo">
-    <button wx:if="{{!hasUserInfo && canIUse}}" open-type="getUserInfo" bindgetuserinfo="getUserInfo"> 获取头像昵称 </button>
-    <block wx:else>
-      <image bindtap="bindViewTap" class="userinfo-avatar" src="{{userInfo.avatarUrl}}" background-size="cover"></image>
-      <text class="userinfo-nickname">{{userInfo.nickName}}</text>
-    </block>
-  </view>
-  <view class="usermotto">
-    <text class="user-motto">{{motto}}</text>
-  </view>
-
-</view>
--->
-
-<view class="container">
-  <view class="user-info">
-      <open-data wx:if="{{canUseopendata}}" class="userinfo-avatar" type="userAvatarUrl"></open-data>
-      <open-data wx:if="{{canUseopendata}}"  class="userinfo-nickname" type="userNickName"></open-data>
-      <image wx:if="{{!canUseopendata}}" class="userinfo-avatar" src="./default.jpg" />
-  </view>
-  <view class="login">
-    <view class="teacherLogin {{teacherFoucs}}" bindtap="touch"  data-ts="teacher" > 
-        <button open-type="getUserInfo" bindgetuserinfo="{{teacherTap}}"  data-ts="teacher" class="login-button button button-glow button-rounded button-border button-caution">教师登录</button>
-    </view>
-    <view class="studentLogin {{studentFoucs}}" bindtap="touch"  data-ts="student">
-        <button open-type="getUserInfo" bindgetuserinfo="{{studentTap}}"  data-ts="student" class="button button-glow button-border button-rounded button-primary login-button login-button-student ">学生登录</button>
-    </view>
-  
-  </view>
-</view>
