+++ client/pages/regis/regis.json
@@ -1,0 +1,1 @@
+{}
