--- client/pages/stu_regis/stu_regis.wxml
@@ -1,13 +1,0 @@
-<view class="container">
-    <text class="title">学生注册页面</text> 
-    <form bindsubmit="regis" bindreset="">
-        <label for="name">学生的姓名</label>
-        <input name="name" id="name" type="text" placeholder="" auto-focus  />
-        <label for="id">学生的学号</label>
-        <input name="id" id="id" type="text" placeholder="" auto-focus  />
-        <label for="college">学生的学院</label>
-        <input name="college" id="college" type="text" placeholder="" auto-focus  />
-        <button form-type="submit">提交</button>
-        <button bindtap="backIndex">返回首页</button>
-    </form>
-</view>
