--- client/pages/stu_regis/stu_regis.wxss
@@ -1,16 +1,0 @@
-.title{
-    display: block;
-}
-form{
-    display: inline-flex;
-    align-items: center;
-    justify-content: center
-}
-form label{
-    display: inline-block;
-}
-#name{
-    display: inline-block;
-    background: rgb(186, 185, 184);
-    vertical-align: middle
-}
