--- client/pages/tea_index/tea_index.wxml
@@ -1,3 +1,0 @@
-<view>
-    教师index界面...待编辑
-</view>
